Construction and setRange tests for ADXL345
The system will test the construction of our ADXL345 using accelerometer
class. It will also test the `setRange` and `getRange` functions using
the newly added Exception classes specific for the ADXL345 accelerometer
class.

The playground now uses the new constant as set by the specification. A
calibration function might be on the cards.

refs #21