Debug from the accelerometer removed
For debugging I created a radian2() function inside the
AccelerometerADXL345 class. However I forgot to remove this.

The playground script now uses better formatting to help with reading
the continuous values from the sensor.

The tests have segmented UpdateAndRead to use their own functions for
scale type. It helps to debug things when they go wrong as we know what
scale we are running at.